- Fixed windows warnings in VELO related classes and HitBase
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@17171 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: b47a8aad0af29d8a032a9ae9d7b2de53eb0e7b4f

+++ LHCb/Tf/TfKernel/TfKernel/DefaultVeloHitManager.h
@@ -1,4 +1,4 @@
-// $Id: DefaultVeloHitManager.h,v 1.8 2007-08-28 12:03:58 jonrob Exp $
+// $Id: DefaultVeloHitManager.h,v 1.9 2007-09-19 13:48:19 krinnert Exp $
 #ifndef INCLUDE_TF_DEFAULTVELOHITMANAGER_H
 #define INCLUDE_TF_DEFAULTVELOHITMANAGER_H 1
 
@@ -129,7 +129,7 @@ namespace Tf {
 
       for(iClus = clusters->begin(); iClus != clusters->end(); ++iClus) {
 
-        double signal = (*iClus)->totalCharge();
+        float signal = static_cast<float>((*iClus)->totalCharge());
 
         // don't use this cluster if we run in ingore mode and the total charge is too low
         if ( m_ignoreLowClusters && (signal < m_lowThreshold) ) continue;
